This commit fixes #4
Script modify_cclm_restart_file.py uses laf...nc files to get a corrected land mask.
Since these files currently only exist for the beginning of the month,
the file could not be found when the run should start in the
middle of the month.

Fix: Do not take the actual starting date but rather the beginning of the
month of the starting date.

This solution might be improved in the future.

+++ scripts/run/modify_cclm_restart_file.py
@@ -8,7 +8,8 @@ def modify_cclm_restart_file(  work_directory_root, # /path/to/work/directory fo
     
     nc_file = work_directory_root + "/" + my_directory + "/mappings/remap_t_grid_exchangegrid_to_" + my_directory + ".nc"
 
-    print('reading in mapping from ' + nc_file)
+    print('Get correct land mask from restart file...')
+    print(' reading in mapping from ' + nc_file)
     nc = netCDF4.Dataset(nc_file, 'r')
     # dst_address starts at 1, python use start with 0
     dst_address = nc.variables['dst_address'][:] - 1 
@@ -16,13 +17,6 @@ def modify_cclm_restart_file(  work_directory_root, # /path/to/work/directory fo
     dst_grid_dims = nc.variables['dst_grid_dims'][:]
     nc.close()
 
-    print('dst_address = ')
-    print(dst_address)
-    print("length = ", len(dst_address))
-    print('src_grid_dims = ')
-    print(src_grid_dims)
-    print('dst_grid_dims = ')
-    print(dst_grid_dims)
 
     # dst_adress = j * I_MAX + i
     I_MAX = dst_grid_dims[0]
@@ -30,31 +24,20 @@ def modify_cclm_restart_file(  work_directory_root, # /path/to/work/directory fo
     j = dst_address // I_MAX
 
     J_MAX = dst_grid_dims[1]
-    #j = dst_address % J_MAX
-    #i = dst_address // J_MAX
-
-    print('i = ')
-    print(i)
-    print("length = ", len(i))
-    print('j = ')
-    print(j)
-    print("length = ", len(j))
     
     nc_dir = work_directory_root + "/" + my_directory + "/OBC"
     nc_file = nc_dir + "/laf" + start_date + "00.nc"
+    
+    if(not os.path.isfile(nc_file)):
+        print(' Error: ' + nc_file + ' does not exist. Abort.')
+        exit()
+        
     os.system("cp --remove-destination `readlink " + nc_file + "` " + nc_file)
 
-    print('reading in restart file ' + nc_file)
+    print(' reading in restart file ' + nc_file)
     nc = netCDF4.Dataset(nc_file, 'r+')
 
     FR_LAND = nc.variables['FR_LAND'][:,:,:] 
-
-    print('FR_LAND')
-    print(FR_LAND)
-    print('')
-    print("length1 = ", len(FR_LAND))
-    print("length2 = ", len(FR_LAND[0]))
-    print("length3 = ", len(FR_LAND[0][0]))
     
     for idx,ii in enumerate(i):
         jj = j[idx]
@@ -64,15 +47,13 @@ def modify_cclm_restart_file(  work_directory_root, # /path/to/work/directory fo
     nc.close()
 
     nc_file = nc_dir + "/laf_modified.nc"
+    print(' write out corrected land mask to ' + nc_file)
+    
     nc = netCDF4.Dataset(nc_file,'w')
     nc.createDimension("xaxis"   ,I_MAX  ); xaxis_var    = nc.createVariable("xaxis"   ,"i4",("xaxis"   ,)); xaxis_var[:]    = [n+1 for n in range(I_MAX)]
     nc.createDimension("yaxis"   ,J_MAX  ); yaxis_var    = nc.createVariable("yaxis"   ,"i4",("yaxis"   ,)); yaxis_var[:]    = [n+1 for n in range(J_MAX)]
     frland = nc.createVariable("FR_LAND"       ,"f8",("yaxis","xaxis",        ))
     frland[:,:] = FR_LAND[0, :, :]
     nc.close()
-
-    #os.system("ln -s " + nc_file + " " + nc_dir + "/laf_modified.nc")
-
-
-
-
+    
+    print('...done')
